+++ auth-api-non-gms/src/main/java/com/omh/android/auth/nongms/domain/auth/AuthUseCase.kt
@@ -16,10 +16,10 @@
 
 package com.omh.android.auth.nongms.domain.auth
 
+import com.omh.android.auth.api.domain.utils.Pkce
+import com.omh.android.auth.api.domain.utils.PkceImpl
 import com.omh.android.auth.nongms.domain.models.ApiResult
 import com.omh.android.auth.nongms.domain.models.OAuthTokens
-import com.omh.android.auth.nongms.domain.utils.Pkce
-import com.omh.android.auth.nongms.domain.utils.PkceImpl
 
 internal class AuthUseCase(
     private val authRepository: AuthRepository,
