+++ auth-api-non-gms/src/test/java/com/omh/android/auth/nongms/usecases/AuthUseCaseTest.kt
@@ -16,11 +16,11 @@
 
 package com.omh.android.auth.nongms.usecases
 
+import com.omh.android.auth.api.domain.utils.Pkce
 import com.omh.android.auth.nongms.domain.auth.AuthRepository
 import com.omh.android.auth.nongms.domain.auth.AuthUseCase
 import com.omh.android.auth.nongms.domain.models.ApiResult
 import com.omh.android.auth.nongms.domain.models.OAuthTokens
-import com.omh.android.auth.nongms.domain.utils.Pkce
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.every
